+++ nf_core/modules/update.py
@@ -6,8 +6,9 @@ from pathlib import Path
 
 import questionary
 
-import nf_core.modules.module_utils
+import nf_core.modules.modules_utils
 import nf_core.utils
+from nf_core.components.components_utils import prompt_component_version_sha
 from nf_core.utils import plural_es, plural_s, plural_y
 
 from .modules_command import ModuleCommand
@@ -151,11 +152,11 @@ class ModuleUpdate(ModuleCommand):
             if sha is not None:
                 version = sha
             elif self.prompt:
-                version = nf_core.modules.module_utils.prompt_module_version_sha(
-                    module, modules_repo=modules_repo, installed_sha=current_version
+                version = nf_core.modules.modules_utils.prompt_component_version_sha(
+                    module, "modules", modules_repo=modules_repo, installed_sha=current_version
                 )
             else:
-                version = modules_repo.get_latest_module_version(module)
+                version = modules_repo.get_latest_component_version(module, self.component_type)
 
             if current_version is not None and not self.force:
                 if current_version == version:
@@ -166,7 +167,7 @@ class ModuleUpdate(ModuleCommand):
                     continue
 
             # Download module files
-            if not self.install_module_files(module, version, modules_repo, install_tmp_dir):
+            if not self.install_component_files(module, version, modules_repo, install_tmp_dir):
                 exit_value = False
                 continue
 
@@ -287,7 +288,7 @@ class ModuleUpdate(ModuleCommand):
             raise LookupError(f"Module '{module}' is not installed in pipeline and could therefore not be updated")
 
         # Check that the supplied name is an available module
-        if module and module not in self.modules_repo.get_avail_modules():
+        if module and module not in self.modules_repo.get_avail_components(self.component_type):
             raise LookupError(
                 f"Module '{module}' not found in list of available modules."
                 f"Use the command 'nf-core modules list remote' to view available software"
@@ -314,7 +315,9 @@ class ModuleUpdate(ModuleCommand):
                 log.info(f"Updating module to ({sha})")
 
         # Check if the update branch is the same as the installation branch
-        current_branch = self.modules_json.get_module_branch(module, self.modules_repo.remote_url, install_dir)
+        current_branch = self.modules_json.get_component_branch(
+            self.component_type, module, self.modules_repo.remote_url, install_dir
+        )
         new_branch = self.modules_repo.branch
         if current_branch != new_branch:
             log.warning(
@@ -360,11 +363,23 @@ class ModuleUpdate(ModuleCommand):
                 for module_dir, module in modules:
                     try:
                         modules_info[repo_name][module_dir].append(
-                            (module, self.sha, self.modules_json.get_module_branch(module, repo_name, module_dir))
+                            (
+                                module,
+                                self.sha,
+                                self.modules_json.get_component_branch(
+                                    self.component_type, module, repo_name, module_dir
+                                ),
+                            )
                         )
                     except KeyError:
                         modules_info[repo_name][module_dir] = [
-                            (module, self.sha, self.modules_json.get_module_branch(module, repo_name, module_dir))
+                            (
+                                module,
+                                self.sha,
+                                self.modules_json.get_component_branch(
+                                    self.component_type, module, repo_name, module_dir
+                                ),
+                            )
                         ]
             elif isinstance(self.update_config[repo_name], dict):
                 # If it is a dict, then there are entries for individual modules or module directories
@@ -380,7 +395,9 @@ class ModuleUpdate(ModuleCommand):
                                         (
                                             module,
                                             custom_sha,
-                                            self.modules_json.get_module_branch(module, repo_name, module_dir),
+                                            self.modules_json.get_component_branch(
+                                                self.component_type, module, repo_name, module_dir
+                                            ),
                                         )
                                     )
                                 except KeyError:
@@ -388,7 +405,9 @@ class ModuleUpdate(ModuleCommand):
                                         (
                                             module,
                                             custom_sha,
-                                            self.modules_json.get_module_branch(module, repo_name, module_dir),
+                                            self.modules_json.get_component_branch(
+                                                self.component_type, module, repo_name, module_dir
+                                            ),
                                         )
                                     ]
                         if self.sha is not None:
@@ -408,7 +427,9 @@ class ModuleUpdate(ModuleCommand):
                                         (
                                             module,
                                             self.sha,
-                                            self.modules_json.get_module_branch(module, repo_name, module_dir),
+                                            self.modules_json.get_component_branch(
+                                                self.component_type, module, repo_name, module_dir
+                                            ),
                                         )
                                     )
                                 except KeyError:
@@ -416,7 +437,9 @@ class ModuleUpdate(ModuleCommand):
                                         (
                                             module,
                                             self.sha,
-                                            self.modules_json.get_module_branch(module, repo_name, module_dir),
+                                            self.modules_json.get_component_branch(
+                                                self.component_type, module, repo_name, module_dir
+                                            ),
                                         )
                                     ]
                             elif isinstance(dir_config[module], str):
@@ -427,7 +450,9 @@ class ModuleUpdate(ModuleCommand):
                                         (
                                             module,
                                             custom_sha,
-                                            self.modules_json.get_module_branch(module, repo_name, module_dir),
+                                            self.modules_json.get_component_branch(
+                                                self.component_type, module, repo_name, module_dir
+                                            ),
                                         )
                                     )
                                 except KeyError:
@@ -435,7 +460,9 @@ class ModuleUpdate(ModuleCommand):
                                         (
                                             module,
                                             custom_sha,
-                                            self.modules_json.get_module_branch(module, repo_name, module_dir),
+                                            self.modules_json.get_component_branch(
+                                                self.component_type, module, repo_name, module_dir
+                                            ),
                                         )
                                     ]
                                 if self.sha is not None:
@@ -454,11 +481,23 @@ class ModuleUpdate(ModuleCommand):
                 for module_dir, module in modules:
                     try:
                         modules_info[repo_name][module_dir].append(
-                            (module, custom_sha, self.modules_json.get_module_branch(module, repo_name, module_dir))
+                            (
+                                module,
+                                custom_sha,
+                                self.modules_json.get_component_branch(
+                                    self.component_type, module, repo_name, module_dir
+                                ),
+                            )
                         )
                     except KeyError:
                         modules_info[repo_name][module_dir] = [
-                            (module, custom_sha, self.modules_json.get_module_branch(module, repo_name, module_dir))
+                            (
+                                module,
+                                custom_sha,
+                                self.modules_json.get_component_branch(
+                                    self.component_type, module, repo_name, module_dir
+                                ),
+                            )
                         ]
                 if self.sha is not None:
                     overridden_repos.append(repo_name)
@@ -518,7 +557,7 @@ class ModuleUpdate(ModuleCommand):
         i = 0
         while i < len(modules_info):
             repo, module, sha = modules_info[i]
-            if not repo.module_exists(module):
+            if not repo.component_exists(module, self.component_type):
                 log.warning(f"Module '{module}' does not exist in '{repo.remote_url}'. Skipping...")
                 modules_info.pop(i)
             elif sha is not None and not repo.sha_exists_on_branch(sha):
@@ -580,7 +619,7 @@ class ModuleUpdate(ModuleCommand):
         pipeline_path = os.path.join(self.dir, "modules", repo_path, module)
 
         log.debug(f"Removing old version of module '{module}'")
-        self.clear_module_dir(module, pipeline_path)
+        self.clear_component_dir(module, pipeline_path)
 
         os.makedirs(pipeline_path)
         for file in files: