+++ nf_core/subworkflows/__init__.py
@@ -1,3 +1,5 @@
 from .create import SubworkflowCreate
 from .install import SubworkflowInstall
+from .list import SubworkflowList
+from .subworkflows_test import SubworkflowsTest
 from .test_yml_builder import SubworkflowTestYmlBuilder