+++ tests/modules/install.py
@@ -22,7 +22,8 @@ def test_modules_install_nopipeline(self):
 @with_temporary_folder
 def test_modules_install_emptypipeline(self, tmpdir):
     """Test installing a module - empty dir given"""
-    self.mods_install.dir = tmpdir
+    os.mkdir(os.path.join(tmpdir, "nf-core-pipe"))
+    self.mods_install.dir = os.path.join(tmpdir, "nf-core-pipe")
     with pytest.raises(UserWarning) as excinfo:
         self.mods_install.install("foo")
     assert "Could not find a 'main.nf' or 'nextflow.config' file" in str(excinfo.value)
@@ -66,4 +67,7 @@ def test_modules_install_different_branch_succeed(self):
 
     # Verify that the branch entry was added correctly
     modules_json = ModulesJson(self.pipeline_dir)
-    assert modules_json.get_module_branch("fastp", GITLAB_URL, GITLAB_REPO) == GITLAB_BRANCH_TEST_BRANCH
+    assert (
+        modules_json.get_component_branch(self.component_type, "fastp", GITLAB_URL, GITLAB_REPO)
+        == GITLAB_BRANCH_TEST_BRANCH
+    )