+++ tests/utils.py
@@ -15,6 +15,7 @@ OLD_TRIMGALORE_BRANCH = "mimic-old-trimgalore"
 GITLAB_URL = "https://gitlab.com/nf-core/modules-test.git"
 GITLAB_REPO = "nf-core"
 GITLAB_DEFAULT_BRANCH = "main-restructure"
+GITLAB_SUBWORKFLOWS_BRANCH = "subworkflows"
 # Branch test stuff
 GITLAB_BRANCH_TEST_BRANCH = "branch-tester-restructure"
 GITLAB_BRANCH_TEST_OLD_SHA = "bce3f17980b8d1beae5e917cfd3c65c0c69e04b5"
@@ -66,3 +67,22 @@ def set_wd(path: Path):
         yield
     finally:
         os.chdir(start_wd)
+
+
+def mock_api_calls(mock, module, version):
+    """Mock biocontainers and anaconda api calls for module"""
+    biocontainers_api_url = f"https://api.biocontainers.pro/ga4gh/trs/v2/tools/{module}/versions/{module}-{version}"
+    anaconda_api_url = f"https://api.anaconda.org/package/bioconda/{module}"
+    mock.register_uri("GET", biocontainers_api_url, text="to modify when the api works and I can know what to add")
+    anaconda_mock = {
+        "status_code": 200,
+        "latest_version": version,
+        "summary": "",
+        "doc_url": "",
+        "dev_url": "",
+        "files": [{"version": version}],
+        "license": "",
+    }
+    biocontainers_mock = {"status_code": 200, "images": [{"image_type": "Docker", "image_name": f"{module}-{version}"}]}
+    mock.register_uri("GET", anaconda_api_url, json=anaconda_mock)
+    mock.register_uri("GET", biocontainers_api_url, json=biocontainers_mock)